Added looping for listener

+++ stt.py
@@ -3,6 +3,7 @@ import whisper
 import ollama_input
 import os
 import json
+import time
 
 import helpers.concat_files as concat
 
@@ -21,19 +22,41 @@ if not os.path.exists(history_folder):
   os.mkdir('./mocchan')
 file_name = f"./mocchan/data.json"
 
-# Records the voice of the user and transcribes it
-record_voice.record()
-transcribed_text = whisper.transcribe()
+counter = 0
 
-# Calls Deepseek to respond to transcribed text
-ollama_response = ollama_input.ollama_chat(message_data, transcribed_text)
+while True:
+  # Records the voice of the user and transcribes it
+  record_voice.record()
+  transcribed_text = whisper.transcribe()
+  
+  if transcribed_text:
+    print(transcribed_text)
 
-# Concats the user's transcribed text and the ollamas
-concat.concat_text(file_name, "user", transcribed_text)
-concat.concat_text(file_name, "assistant", ollama_response["answer"])
+    # Calls Deepseek to respond to transcribed text
+    ollama_response = ollama_input.ollama_chat(message_data, transcribed_text)
 
-print(f"\033[96m[YOU]: {transcribed_text}") # For Cyan colored text: \033[96m
-print(f"\033[92m[MOCCHAN]: " + ollama_response["answer"]) # For Green colored text: \033[92m
+    # Concats the user's transcribed text and the ollamas
+    concat.concat_text(file_name, "user", transcribed_text)
+    concat.concat_text(file_name, "assistant", ollama_response["answer"])
 
-# Removes audio file
-os.remove("./audio.wav")
+    print("")
+    print(f"\033[96m[YOU]: {transcribed_text}") # For Cyan colored text: \033[96m
+    print(f"\033[92m[MOCCHAN]: " + ollama_response["answer"]) # For Green colored text: \033[92m
+    print("\033[37m") # Resets the color of text back to white
+
+    # Removes audio file
+    os.remove("./audio.wav")
+
+    time.sleep(3)
+  else:
+    print("NO TEXT!!")
+
+    if counter == 5:
+      print("Ending Listener")
+      break
+    else:
+      counter+=1
+      print(f"Keeing an ear out. Retrying {6 - counter} more times.")
+
+
+
